dff_api/requirements.txt
aiohttp==3.9.4
pillow==10.3.0
dff[telegram]